move more files around
