+++ local/.env.example
@@ -1,0 +1,4 @@
+HUGGINGFACE_HUB_CACHE=""
+MONGODB_URI=""
+MONGODB_DB=""
+MONGODB_COLLECTION_LOCAL=""
