+++ local/gguf/llm_bench_gguf/create_model.py
@@ -11,12 +11,8 @@ OUTPUT_DIR = "/gemini/gguf"
 
 
 def main() -> None:
-    parser = argparse.ArgumentParser(
-        description="Download a model from the Hugging Face Hub."
-    )
-    parser.add_argument(
-        "-m", "--model", required=True, help="Model ID from the Hugging Face Hub."
-    )
+    parser = argparse.ArgumentParser(description="Download a model from the Hugging Face Hub.")
+    parser.add_argument("-m", "--model", required=True, help="Model ID from the Hugging Face Hub.")
     args = parser.parse_args()
 
     # Clean the model id
