+++ local/huggingface/llm_bench_hf/server.py
@@ -15,6 +15,9 @@ from llm_bench_api.logging import log_to_mongo
 from llm_bench_api.utils import check_and_clean_space
 from llm_bench_api.utils import has_existing_run
 
+from llm_bench_hf.tgi import generate as generate_tgi
+from llm_bench_hf.transformers import generate as generate_transformers
+
 log_path = "/var/log/llm_benchmarks.log"
 try:
     logging.basicConfig(filename=log_path, level=logging.INFO)
@@ -89,7 +92,15 @@ def call_huggingface() -> Union[Response, Tuple[Response, int]]:
                 logger.info("Re-running benchmark anyway because run_always is True")
             else:
                 logger.info(f"Model has been benchmarked before: {model_name}, quant: {quant_str}")
-                return jsonify({"status": "skipped", "reason": "model has been benchmarked before"}), 200
+                return (
+                    jsonify(
+                        {
+                            "status": "skipped",
+                            "reason": "model has been benchmarked before",
+                        }
+                    ),
+                    200,
+                )
         else:
             logger.info(f"Model has not been benchmarked before: {model_name}, quant: {quant_str}")
 
@@ -97,9 +108,9 @@ def call_huggingface() -> Union[Response, Tuple[Response, int]]:
         check_and_clean_space(directory=CACHE_DIR, threshold=90.0)
 
         if framework == "transformers":
-            from app.transformers import generate
+            generate = generate_transformers
         elif framework == "hf-tgi":
-            from app.tgi import generate
+            generate = generate_tgi
         else:
             raise ValueError(f"Unknown framework: {framework}")
