+++ local/huggingface/pyproject.toml
@@ -13,3 +13,8 @@ torch = "^2.2.2"
 transformers = "^4.40.0"
 flask = "^3.0.3"
 numpy = "^1.26.4"
+click = "^8.1.7"
+pymongo = "^4.6.3"
+pynvml = "^11.5.0"
+pytz = "^2024.1"
+
